requirements.txt
pymavlink==2.4.48
mavsdk==3.8.0
python-dotenv==1.1.1
protobuf==5.29.0
